Title: 배열의 특정 위치 찾기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240930/배열의 특정 위치 찾기/find-specific-location-fo-array.c
@@ -18,7 +18,7 @@ int main() {
             count++;
         }
     }
-    double c = b/count;
+    float c = b/count;
 
     printf("%d %.1f", a,c);
